tests: keep the view half of the styling tests out of the AI stack
lumen/tests/views runs in the core environment, which has no pydantic,
so the tests that build the plot agent's schema move to lumen/tests/ai
alongside the other agent tests.
