+++ lumen/tests/views/test_hvplot_styling.py
@@ -1,12 +1,8 @@
-import json
-
 import pandas as pd
 import pytest
 
 from hvplot.ui import hvDataFrameExplorer
 
-from lumen.ai.agents.hvplot import hvPlotAgent
-from lumen.ai.config import PROMPTS_DIR
 from lumen.pipeline import Pipeline
 from lumen.sources.base import InMemorySource
 from lumen.views.base import (
@@ -125,34 +121,11 @@ def test_style_params_round_trip_through_the_spec(pipeline):
     assert spec["rot"] == 45
 
 
-# ---- What the LLM is handed ----
-
-def test_style_params_reach_the_agent_schema():
-    schema = {"lon": {"type": "number"}, "lat": {"type": "number"}}
-
-    properties = hvPlotAgent()._get_model("main", schema).model_json_schema()["properties"]
-
-    assert set(HVPLOT_STYLE_PARAMS) <= set(properties)
-    for name in HVPLOT_STYLE_PARAMS:
-        assert properties[name].get("description"), name
-
+# ---- The colormap choices ----
 
 def test_the_colormap_enum_stays_short():
     """hvPlot offers 712 colormaps and every one would be spent as prompt."""
     assert len(hvPlotBaseView.param.cmap.objects) <= 20
-
-
-def test_the_schema_stays_affordable():
-    """The whole point of an allowlist is that it does not grow by accident.
-
-    Raise these numbers deliberately, having looked at what was added.
-    """
-    schema = {"lon": {"type": "number"}, "lat": {"type": "number"},
-              "family": {"type": "string"}, "pop": {"type": "number"}}
-
-    rendered = json.dumps(hvPlotAgent()._get_model("main", schema).model_json_schema())
-
-    assert len(rendered) <= 7000
 
 
 def test_a_dict_colormap_is_still_accepted(pipeline, df):
@@ -163,14 +136,6 @@ def test_a_dict_colormap_is_still_accepted(pipeline, df):
     view = hvPlotView(pipeline=pipeline, kind="line", x="x", y="y", cmap=color_key)
 
     assert record_hvplot_call(view, df)["cmap"] == color_key
-
-
-def test_prompt_asks_for_styling_only_when_requested():
-    """The options are worth having only if the model leaves them alone by
-    default, so the prompt has to say so."""
-    prompt = (PROMPTS_DIR / "hvPlotAgent" / "main.jinja2").read_text()
-
-    assert "only when the request asks" in prompt
 
 
 def test_the_view_title_is_not_drawn_twice(pipeline, df):
